+++ include/mxnet-cpp/MxNetCpp.h
@@ -1,0 +1,19 @@
+/*!
+ *  Copyright (c) 2016 by Contributors
+ * \file MxNetCpp.h
+ * \brief meta include file for mxnet.cpp
+ * \author Chuntao Hong, Zhang Chen
+ */
+
+#ifndef MXNETCPP_H_
+#define MXNETCPP_H_
+
+#include "mxnet-cpp/executor.hpp"
+#include "mxnet-cpp/symbol.hpp"
+#include "mxnet-cpp/ndarray.hpp"
+#include "mxnet-cpp/operator.hpp"
+#include "mxnet-cpp/optimizer.hpp"
+#include "mxnet-cpp/kvstore.hpp"
+#include "mxnet-cpp/op.h"
+
+#endif  // MXNETCPP_H_
