+++ include/mxnet-cpp/logging.h
