+++ include/mxnet-cpp/op.h
@@ -10,9 +10,9 @@
 
 #include <string>
 #include <vector>
-#include "mxnet/base.h"
-#include "mxnet/shape.h"
-#include "mxnet/MxNetCpp.h"
+#include "mxnet-cpp/base.h"
+#include "mxnet-cpp/shape.h"
+#include "mxnet-cpp/MxNetCpp.h"
 
 namespace mxnet {
 namespace cpp {
