+++ include/mxnet-cpp/operator.h
@@ -11,9 +11,9 @@
 #include <map>
 #include <string>
 #include <vector>
-#include "mxnet/base.h"
-#include "mxnet/op_map.h"
-#include "mxnet/symbol.h"
+#include "mxnet-cpp/base.h"
+#include "mxnet-cpp/op_map.h"
+#include "mxnet-cpp/symbol.h"
 
 namespace mxnet {
 namespace cpp {
