+++ include/mxnet-cpp/optimizer.hpp
@@ -12,7 +12,7 @@
 #include <map>
 #include <string>
 #include <vector>
-#include "mxnet/optimizer.h"
+#include "mxnet-cpp/optimizer.h"
 
 namespace mxnet {
 namespace cpp {
