--- include/mxnet/MxNetCpp.h
@@ -1,19 +1,0 @@
-/*!
- *  Copyright (c) 2016 by Contributors
- * \file MxNetCpp.h
- * \brief meta include file for mxnet.cpp
- * \author Chuntao Hong, Zhang Chen
- */
-
-#ifndef MXNETCPP_H_
-#define MXNETCPP_H_
-
-#include "mxnet/executor.hpp"
-#include "mxnet/symbol.hpp"
-#include "mxnet/ndarray.hpp"
-#include "mxnet/operator.hpp"
-#include "mxnet/optimizer.hpp"
-#include "mxnet/kvstore.hpp"
-#include "mxnet/op.h"
-
-#endif  // MXNETCPP_H_
